Move classes into files
